SystemUI: Introduce user interface for Alert Sliders.
Ported from OxygenOS and reworked for our alert slider
implementation. We target AudioManager instead of Zen,
icons are also the same as aosp and the dialog uses the
material theme as well as support for our themes.

To use, the alert slider config must be enabled. By default,
the dialog shows on the left side. To move it to the right
side, set the location config to 1.

Change-Id: I0522b849a8aeb3af0936eec4c0def894931cc564
Signed-off-by: klozz <klozz@thexperienceproject.org>